refactor: export scopeStyle method
